+++ CHANGELOG.md
@@ -6,6 +6,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Changed
 * Don't emit error on S3 abort
 * Increase max post size to 10mb
+* Removed node-fs dependency
 
 ### Fixed
 * Emit finish event on local writeStream